+++ packages/stylelint-plugin/src/rules/spiracss-rel-comments.types.ts
@@ -16,6 +16,7 @@ export type Options = {
     }
   }
   fileCase: FileNameCase
+  childFileCase?: FileNameCase
   validate: {
     path: boolean
   }